resource_server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
PyJWT[crypto]>=2.10.0
httpx>=0.27.0
pytest>=8.0.0
